Hardcode GHCup logs to stdout
It turns out that GHCup hardcodes its console output to use
ANSI escape sequences (via 'pretty-terminal'). Thus, for now it seems
useless to output these logs to anything other than a console. It
would be nice to instead redirect these to e.g. a graphical toast, but
for now it seems the GUI needs to handle such things itself.